app/requirements.txt
Flask==2.0.1
python-consul==1.1.0
